Adminconsole commands
can now use adminconsole to place blocks and give blocks to players

+++ RoadsClient/src/GameObjects/World.java
@@ -13,7 +13,7 @@ public class World {
 	private static Block[][] world;
 
 	public static double cameraX, cameraY;
-	public static double playerid = -1;
+	public static int playerid = -1;
 	public static Block[] playerInventory = new Block[5];
 	public static int selectedInventory = 0;
